added qr code generator component
