+++ Corona-Warn-App/src/deviceForTesters/java/de/rki/coronawarnapp/test/eventregistration/ui/qrcode/QrCodeCreationTestViewModel.kt
@@ -13,6 +13,7 @@ import com.google.zxing.common.BitMatrix
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import de.rki.coronawarnapp.appconfig.AppConfigProvider
+import de.rki.coronawarnapp.ui.eventregistration.organizer.details.QrCodeGenerator
 import de.rki.coronawarnapp.util.coroutine.DispatcherProvider
 import de.rki.coronawarnapp.util.di.AppContext
 import de.rki.coronawarnapp.util.files.FileSharing
@@ -26,6 +27,7 @@ import java.io.FileOutputStream
 class QrCodeCreationTestViewModel @AssistedInject constructor(
     private val dispatcher: DispatcherProvider,
     private val fileSharing: FileSharing,
+    private val qrCodeGenerator: QrCodeGenerator,
     @AppContext private val context: Context,
     private val appConfigProvider: AppConfigProvider,
 ) : CWAViewModel(dispatcher) {
@@ -38,7 +40,7 @@ class QrCodeCreationTestViewModel @AssistedInject constructor(
      * Creates a QR Code [Bitmap] ,result is delivered by [qrCodeBitmap]
      */
     fun createQrCode(input: String) = launch(context = dispatcher.IO) {
-        qrCodeBitmap.postValue(encodeAsBitmap(input))
+        qrCodeBitmap.postValue(qrCodeGenerator.createQrCode(input))
     }
 
     /**
@@ -82,48 +84,6 @@ class QrCodeCreationTestViewModel @AssistedInject constructor(
         if (!dir.exists()) dir.mkdirs()
         return File(dir, "CoronaWarnApp-Event.pdf")
     }
-
-    private suspend fun encodeAsBitmap(input: String, size: Int = 1000): Bitmap? {
-        return try {
-            val qrCodeErrorCorrectionLevel = appConfigProvider
-                .getAppConfig()
-                .presenceTracing
-                .qrCodeErrorCorrectionLevel
-            Timber.i("QrCodeErrorCorrectionLevel: $qrCodeErrorCorrectionLevel")
-            val hints = mapOf(
-                EncodeHintType.ERROR_CORRECTION to qrCodeErrorCorrectionLevel
-                // This is not required in the specs and it should not be enabled
-                // it is causing crash on older Android versions ex:API 23
-                // EncodeHintType.CHARACTER_SET to Charsets.UTF_8
-            )
-            MultiFormatWriter().encode(
-                input,
-                BarcodeFormat.QR_CODE,
-                size,
-                size,
-                hints
-            ).toBitmap()
-        } catch (e: Exception) {
-            Timber.d(e, "Qr code creation failed")
-            errorMessage.postValue(e.localizedMessage ?: "QR code creation failed")
-            null
-        }
-    }
-
-    private fun BitMatrix.toBitmap() =
-        Bitmap.createBitmap(
-            context.resources.displayMetrics,
-            width,
-            height,
-            Bitmap.Config.ARGB_8888
-        ).apply {
-            for (x in 0 until width) {
-                for (y in 0 until height) {
-                    val color = if (get(x, y)) BLACK else WHITE
-                    setPixel(x, y, color)
-                }
-            }
-        }
 
     @AssistedFactory
     interface Factory : SimpleCWAViewModelFactory<QrCodeCreationTestViewModel>
